+++ rose-bitset-derive/src/impls/tests.rs
@@ -690,7 +690,130 @@ pub fn generate_code(
                 assert!(SET_D.is_subset(SET_D));
             }
 
-            // todo!("test fn is_strict_subset(Self, Self) -> bool")
+            // fn is_strict_subset(Self, Self) -> bool
+
+            #[test]
+            fn empty_is_strict_subset_of_everything_but_itself() {
+                let empty = #ident::new();
+                assert!(!empty.is_strict_subset(empty));
+                assert!(empty.is_strict_subset(#ident::all()));
+                assert!(empty.is_strict_subset(SET_A));
+                assert!(empty.is_strict_subset(SET_B));
+                assert!(empty.is_strict_subset(SET_C));
+                assert!(empty.is_strict_subset(SET_D));
+            }
+
+            #[test]
+            fn all_isnt_strict_subset_of_anything() {
+                let all = #ident::all();
+                assert!(!all.is_strict_subset(#ident::new()));
+                assert!(!all.is_strict_subset(all));
+                assert!(!all.is_strict_subset(SET_A));
+                assert!(!all.is_strict_subset(SET_B));
+                assert!(!all.is_strict_subset(SET_C));
+                assert!(!all.is_strict_subset(SET_D));
+            }
+
+            #[test]
+            fn a_isnt_strict_subset_of_complement() {
+                assert!(!SET_A.is_strict_subset(SET_A.complement()));
+            }
+
+            #[test]
+            fn a_isnt_strict_subset_of_a() {
+                assert!(!SET_A.is_strict_subset(SET_A));
+            }
+
+            #[test]
+            fn a_isnt_strict_subset_of_b() {
+                assert!(!SET_A.is_strict_subset(SET_B));
+            }
+
+            #[test]
+            fn a_isnt_strict_subset_of_c() {
+                assert!(!SET_A.is_strict_subset(SET_C));
+            }
+
+            #[test]
+            fn a_isnt_strict_subset_of_d() {
+                assert!(!SET_A.is_strict_subset(SET_D));
+            }
+
+            #[test]
+            fn b_isnt_strict_subset_of_complement() {
+                assert!(!SET_B.is_strict_subset(SET_B.complement()));
+            }
+
+            #[test]
+            fn b_isnt_strict_subset_of_a() {
+                assert!(!SET_B.is_strict_subset(SET_A));
+            }
+
+            #[test]
+            fn b_isnt_strict_subset_of_b() {
+                assert!(!SET_B.is_strict_subset(SET_B));
+            }
+
+            #[test]
+            fn b_isnt_strict_subset_of_c() {
+                assert!(!SET_B.is_strict_subset(SET_C));
+            }
+
+            #[test]
+            fn b_isnt_strict_subset_of_d() {
+                assert!(!SET_B.is_strict_subset(SET_D));
+            }
+
+            #[test]
+            fn c_isnt_strict_subset_of_complement() {
+                assert!(!SET_C.is_strict_subset(SET_C.complement()));
+            }
+
+            #[test]
+            fn c_isnt_strict_subset_of_a() {
+                assert!(!SET_C.is_strict_subset(SET_A));
+            }
+
+            #[test]
+            fn c_isnt_strict_subset_of_b() {
+                assert!(!SET_C.is_strict_subset(SET_B));
+            }
+
+            #[test]
+            fn c_isnt_strict_subset_of_c() {
+                assert!(!SET_C.is_strict_subset(SET_C));
+            }
+
+            #[test]
+            fn c_isnt_strict_subset_of_d() {
+                assert!(!SET_C.is_strict_subset(SET_D));
+            }
+
+            #[test]
+            fn d_isnt_strict_subset_of_complement() {
+                assert!(!SET_D.is_strict_subset(SET_D.complement()));
+            }
+
+            #[test]
+            fn d_isnt_strict_subset_of_a() {
+                assert!(!SET_D.is_strict_subset(SET_A));
+            }
+
+            #[test]
+            fn d_isnt_strict_subset_of_b() {
+                assert!(!SET_D.is_strict_subset(SET_B));
+            }
+
+            #[test]
+            fn d_isnt_strict_subset_of_c() {
+                assert!(!SET_D.is_strict_subset(SET_C));
+            }
+
+            #[test]
+            fn d_isnt_strict_subset_of_d() {
+                assert!(!SET_D.is_strict_subset(SET_D));
+            }
+
             // todo!("test fn is_superset(Self, Self) -> bool")
             // todo!("test fn is_strict_superset(Self, Self) -> bool")
             // todo!("test fn len(Self) -> usize")